Improvements made to opcode_reader class
Changed the condition from this->opcode_address != 0 to this->opcode_address != nullptr. This ensures that we're checking against a null pointer using the C++ nullptr keyword, which is more idiomatic and clearer.
Changed the condition from this->opcode_address != 0 to this->opcode_address != nullptr. This ensures that we're checking against a null pointer using the C++ nullptr keyword, which is more idiomatic and clearer.